will show modal only after the account balance is updated
