+++ app/components/UI/DrawerView/index.js
@@ -410,6 +410,8 @@ class DrawerView extends PureComponent {
 		networkSelected: false,
 		networkType: undefined,
 		networkCurrency: undefined,
+		balance: null,
+		balanceUpdated: false,
 	};
 
 	browserSectionRef = React.createRef();
@@ -532,8 +534,7 @@ class DrawerView extends PureComponent {
 	};
 
 	onNetworksModalCloses = async (manualClose) => {
-		this.toggleNetworksModal();
-		this.setState({ networkSelected: !this.state.networkSelected });
+		this.setState({ networkSelected: !this.state.networkSelected, balanceUpdated: false });
 		if (!manualClose) {
 			await this.hideDrawer();
 			await this.setState({ networkSelected: !this.state.networkSelected });
@@ -550,7 +551,16 @@ class DrawerView extends PureComponent {
 		}
 	};
 
+	accountBalanceUpdated = () => {
+		const that = this;
+		setTimeout(() => {
+			that.setState({ balanceUpdated: true });
+		}, 10000);
+	};
+
 	onNetworkSelected = (type, currency) => {
+		this.setState({ balanceUpdated: false });
+		this.accountBalanceUpdated();
 		this.setState({
 			networkType: type,
 			networkCurrency: currency,
@@ -987,6 +997,7 @@ class DrawerView extends PureComponent {
 		};
 		const { name, ens } = account;
 		account.balance = (accounts[selectedAddress] && renderFromWei(accounts[selectedAddress].balance)) || 0;
+		this.setState({ balance: account.balance });
 		const fiatBalance = Engine.getTotalFiatAccountBalance();
 		if (fiatBalance !== this.previousBalance) {
 			this.previousBalance = this.currentBalance;
@@ -994,6 +1005,8 @@ class DrawerView extends PureComponent {
 		this.currentBalance = fiatBalance;
 		const fiatBalanceStr = renderFiat(this.currentBalance, currentCurrency);
 		const accountName = isDefaultAccountName(name) && ens ? ens : name;
+		const showInfoModal =
+			this.state.networkSelected && Number(this.state.balance) === 0 && this.state.balanceUpdated;
 
 		return (
 			<View style={styles.wrapper} testID={'drawer-screen'}>
@@ -1150,7 +1163,22 @@ class DrawerView extends PureComponent {
 					backdropColor={colors.overlay.default}
 					backdropOpacity={1}
 				>
-					{this.state.networkSelected ? (
+					<NetworkList
+						navigation={this.props.navigation}
+						onClose={this.onNetworksModalClose}
+						onNetworkSelected={this.onNetworkSelected}
+						showInvalidCustomNetworkAlert={this.showInvalidCustomNetworkAlert}
+					/>
+				</Modal>
+				{showInfoModal && (
+					<Modal
+						isVisible={showInfoModal}
+						onBackdropPress={this.onNetworksModalCloses}
+						onBackButtonPress={this.onNetworksModalCloses}
+						onSwipeComplete={this.onNetworksModalCloses}
+						swipeDirection={'down'}
+						propagateSwipe
+					>
 						<NetworkInfo
 							onClose={this.onNetworksModalCloses}
 							type={this.state.networkType}
@@ -1163,8 +1191,9 @@ class DrawerView extends PureComponent {
 							onNetworkSelected={this.onNetworkSelected}
 							showInvalidCustomNetworkAlert={this.showInvalidCustomNetworkAlert}
 						/>
-					)}
+					)
 				</Modal>
+			)}
 				<Modal backdropColor={colors.overlay.default} backdropOpacity={1} isVisible={!!invalidCustomNetwork}>
 					<InvalidCustomNetworkAlert
 						navigation={this.props.navigation}
