+++ app/components/UI/NetworkList/index.js
@@ -149,6 +149,7 @@ export class NetworkList extends PureComponent {
 
 	onNetworkChange = (type) => {
 		this.props.onNetworkSelected(type, 'ETH');
+		this.props.onClose(false);
 		const { NetworkController, CurrencyRateController } = Engine.context;
 		CurrencyRateController.setNativeCurrency('ETH');
 		NetworkController.setProviderType(type);
@@ -202,6 +203,8 @@ export class NetworkList extends PureComponent {
 			block_explorer_url: blockExplorerUrl,
 			network_name: 'rpc',
 		});
+
+		this.props.onClose(false);
 	};
 
 	getStyles = () => {
